% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_utility.R
\name{clean_transaction_export}
\alias{clean_transaction_export}
\title{Cleans the raw Vanguard transaction history.}
\usage{
clean_transaction_export(filename)
}
\arguments{
\item{filename}{A string corresponding to the filepath of the
exported Vanguard data (.xlsx format).}
}
\value{
The cleaned transaction history tibble
}
\description{
Vanguard permits a user to download the transaction history for one or more
of the accounts that user owns. This function cleans up this file by removing
redundant information and calculating cumulative quantities over time.
To do so, this function removes the cash holdings and transactions between
banks and the settlement fund--only transactions of mutual funds
(excluding money market accounts), ETFs, stocks, and bonds
are tracked. Dividends and capital gains that get reinvested are also
retained. But the raw transaction history records both the accumulation of the
dividend and its reinvestment, so we retain only the reinvestment
(recorded with a minus sign, like is done when buying a share of a stock).

No edits should be made to the contents of the excel file Vanguard exports.
If it is in .csv format, convert to .xlsx first.
}
\examples{
\dontrun{}
# Vanguard exports the file 'ofxdownload.csv' which we save as
# 'raw_data.xlsx.'

clean_transaction_export('ofxdownload.xlsx')
}
