% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processraw.R
\name{complete_history}
\alias{complete_history}
\title{Generate complete historical data from cleaned transaction history.}
\usage{
complete_history(cleaned_raw_df)
}
\arguments{
\item{cleaned_raw_df}{A cleaned transaction history tibble.}
}
\value{
Complete portfolio history
}
\description{
Once the exported Vanguard data has been initially cleaned with
\code{\link[=clean_transaction_export]{clean_transaction_export()}}, we must interpolate the
holdings for all days, even if no transactions were recorded.
This function forms the complete time series for each stock and
Vanguard account. This will allow us to calculate the daily
portfolio balance once we merge in price data.
}
\examples{
\dontrun{}
# Vanguard exports the file 'ofxdownload.csv' which we save as
# 'raw_data.xlsx.'

cleaned_df = clean_transaction_export('raw_data.xlsx')
complete_df = complete_history(cleaned_df)
}
