% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_portfolio}
\alias{generate_portfolio}
\title{Generate all output tibbles produced from raw import data.}
\usage{
generate_portfolio(input_filename)
}
\arguments{
\item{input_filename}{The path to the exported Vanguard data}
}
\value{
A list of tibbles used for analysis and plotting.
\itemize{
\item processed_raw - A cleaned tibble of all transactions.
\item balance_df - A day level history of each holding's share count and
cost basis.
\item price_history - A daily history of each individual holding's share count,
current value, and cost_basis.
\item portfolio_history - A portfolio-level daily balance history.
}
}
\description{
Runs all the processing scripts at once for quick data generation.
}
