% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cleaned_history.R
\name{prepare_price_history}
\alias{prepare_price_history}
\title{Merge in price data into complete portfolio history.}
\usage{
prepare_price_history(input_balance_df)
}
\arguments{
\item{input_balance_df}{The complete balance_history df.}
}
\value{
The full price history of each holdings in the portfolios.
}
\description{
Once the user has generated the complete balance history using
\code{\link[=clean_transaction_export]{clean_transaction_export()}} followed by \code{\link[=complete_history]{complete_history()}},
we obtain a dataset with the holdings for every single day between
the first and last transaction in the raw dataset.

This function scrapes the price of each stock using the BatchGetSymbols
library and then computes the balance value for each ticker in each account.
}
\examples{
\dontrun{}
# Vanguard exports the file 'ofxdownload.csv' which we save as
# 'raw_data.xlsx.'

cleaned_df = clean_transaction_export('raw_data.xlsx')
complete_df = complete_history(cleaned_df)

price_history_df = prepare_price_history(complete_df)
}
