% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cleaned_history.R
\name{prepare_return_data}
\alias{prepare_return_data}
\title{Compute portfolio wide performance history for each account}
\usage{
prepare_return_data(processed_price_df)
}
\arguments{
\item{processed_price_df}{The tibble of complete price history}
}
\value{
The full portfolio-level history of for each account.
}
\description{
Once the user has generated the complete price history
for each holding in each account, using \code{\link[=prepare_price_history]{prepare_price_history()}},
this function aggregates each day's holdings to compute total
portfolio value and cost basis over time.
}
\examples{
\dontrun{}
# Vanguard exports the file 'ofxdownload.csv' which we save as
# 'raw_data.xlsx.'

cleaned_df = clean_transaction_export('raw_data.xlsx')
complete_df = complete_history(cleaned_df)

price_history_df = prepare_price_history(complete_df)
prepare_return_data(price_history_df)
}
